refactor: consolidate duplicate UI components and remove dead code
- Inline page-new.tsx content into main pages (home, garden, chat)
- Remove isNewUIEnabled() feature flag checks from consolidated pages
- Rename *New.tsx components to clean names (PartsList, CheckInSlots, WeekSelector)
- Delete old component implementations that were behind always-true feature flag
- Remove lib/inbox/pragmaticData.ts (unused fallback data)

Net reduction: ~1,143 lines of duplicate/dead code
