+++ app/(tabs)/garden/page.tsx
@@ -4,21 +4,16 @@ import { useState, useEffect } from 'react'
 import { searchPartsV2 } from '@/lib/data/parts-lite'
 import type { PartRowV2 } from '@/lib/data/schema/types'
 import { PartsList } from '@/components/garden/PartsList'
-import { PageContainer } from '@/components/common/PageContainer'
-import { isNewUIEnabled } from '@/config/features'
-import { GardenPageNew } from './page-new'
+import { MaterialIcon } from '@/components/ui/MaterialIcon'
+import Link from 'next/link'
 
 export default function GardenPage() {
-  const newUI = isNewUIEnabled()
-  
-  // Always call hooks - they must be called unconditionally
   const [parts, setParts] = useState<PartRowV2[]>([])
   const [error, setError] = useState<string | null>(null)
   const [isLoading, setIsLoading] = useState(true)
+  const [searchQuery, setSearchQuery] = useState('')
 
   useEffect(() => {
-    // Only fetch if using old UI
-    if (newUI) return
     let isActive = true
 
     async function fetchPartsData() {
@@ -49,73 +44,77 @@ export default function GardenPage() {
     return () => {
       isActive = false
     }
-  }, [newUI])
+  }, [])
 
-  if (newUI) {
-    return <GardenPageNew />
-  }
-
-  const totalParts = parts.length
-  const establishedCount = parts.filter(
-    (p) => p.status === 'acknowledged' || p.status === 'active' || p.status === 'integrated'
-  ).length
-  const activeToday = parts.filter((p) => {
-    if (!p.last_active) return false
-    const hoursAgo = (Date.now() - new Date(p.last_active).getTime()) / (1000 * 60 * 60)
-    return hoursAgo < 24
-  }).length
-  const avgConfidence = totalParts > 0 
-    ? Math.round((parts.reduce((sum, p) => sum + p.confidence, 0) / totalParts) * 100)
-    : 0
+  const filteredParts = searchQuery
+    ? parts.filter((part) =>
+        part.name?.toLowerCase().includes(searchQuery.toLowerCase())
+      )
+    : parts
 
   return (
-    <div className="min-h-screen flex flex-col">
-      <PageContainer className="flex-1 py-6 md:py-8 space-y-8">
-        <header>
-          <h1 className="text-4xl font-bold tracking-tight">Your Inner Garden</h1>
-          <p className="text-muted-foreground mt-2">
-            {isLoading ? '—' : totalParts} parts discovered • {isLoading ? '—' : activeToday} active today
-          </p>
+    <div className="min-h-screen bg-[var(--hs-bg)] flex flex-col hs-animate-in">
+      {/* Header */}
+      <header className="pt-12 pb-4 px-5">
+        <div className="flex items-center justify-between">
+          <div className="flex-1">
+            <h1 className="text-[28px] font-bold text-[var(--hs-text-primary)] leading-tight">
+              My Parts
+            </h1>
+            <p className="text-sm text-[var(--hs-text-secondary)] mt-1">
+              {parts.length} {parts.length === 1 ? 'part' : 'parts'} in your garden
+            </p>
+          </div>
+        </div>
+      </header>
 
-          {/* Stats Banner */}
-          {!isLoading && (
-            <div
-              className="mt-6 rounded-lg border border-border/40 bg-card/20 p-5 backdrop-blur grid grid-cols-3 gap-4 md:gap-6"
-              aria-live="polite"
-            >
-              <div>
-                <p className="text-xs font-medium uppercase text-muted-foreground tracking-wider">
-                  Total Parts
-                </p>
-                <p className="text-2xl md:text-3xl font-bold mt-2">{totalParts}</p>
-              </div>
-              <div>
-                <p className="text-xs font-medium uppercase text-muted-foreground tracking-wider">
-                  Established
-                </p>
-                <p className="text-2xl md:text-3xl font-bold mt-2">{establishedCount}</p>
-                <p className="text-xs text-muted-foreground/70 mt-1">avg {avgConfidence}% sure</p>
-              </div>
-              <div>
-                <p className="text-xs font-medium uppercase text-muted-foreground tracking-wider">
-                  Active Today
-                </p>
-                <p className="text-2xl md:text-3xl font-bold mt-2">{activeToday}</p>
-              </div>
-            </div>
-          )}
-        </header>
+      {/* Search Bar */}
+      <div className="px-5 pb-4">
+        <div className="relative">
+          <div className="absolute inset-y-0 left-0 pl-4 flex items-center pointer-events-none">
+            <MaterialIcon
+              name="search"
+              className="text-xl text-[var(--hs-text-tertiary)]"
+            />
+          </div>
+          <input
+            type="search"
+            className="hs-input pl-12 pr-4"
+            placeholder="Search parts..."
+            value={searchQuery}
+            onChange={(e) => setSearchQuery(e.target.value)}
+          />
+        </div>
+      </div>
 
-        <main className="flex-grow">
-          {error && (
-            <div className="text-red-500 text-center p-6 rounded-lg border border-red-500/20 bg-red-500/5">
-              <p>Could not load garden: {error}</p>
-            </div>
-          )}
+      {/* Parts Grid */}
+      <main className="flex-1 px-5 pb-24">
+        {error && (
+          <div className="hs-card p-6 text-center border border-red-200 dark:border-red-900/30 bg-red-50 dark:bg-red-900/10">
+            <MaterialIcon name="error" className="text-3xl text-red-500 mb-2" />
+            <p className="text-sm text-red-600 dark:text-red-400">
+              Could not load garden: {error}
+            </p>
+          </div>
+        )}
 
-          {!error && <PartsList parts={parts} isLoading={isLoading} />}
-        </main>
-      </PageContainer>
+        {!error && (
+          <div className="grid grid-cols-2 gap-4">
+            <PartsList parts={filteredParts} isLoading={isLoading} />
+          </div>
+        )}
+      </main>
+
+      {/* Floating Action Button */}
+      <div className="fixed bottom-24 right-5 z-20">
+        <Link
+          href="/garden/new"
+          className="flex items-center justify-center w-14 h-14 rounded-full bg-[var(--hs-primary)] text-white shadow-glow hover:bg-[var(--hs-primary-dark)] transition-all duration-200 hover:scale-105"
+          aria-label="Add new part"
+        >
+          <MaterialIcon name="add" className="text-3xl" />
+        </Link>
+      </div>
     </div>
   )
 }
