+++ app/chat/page.tsx
@@ -2,41 +2,56 @@
 
 import { Suspense, useEffect, useMemo, useState } from "react"
 import Link from "next/link"
-import { Inter } from "next/font/google"
 import { EtherealChat } from "@/components/ethereal/EtherealChat"
-import { isNewUIEnabled } from "@/config/features"
-import { ChatPageNew } from "./page-new"
-
-const inter = Inter({ subsets: ["latin"], weight: ["100", "300", "400", "600"], variable: "--font-ethereal" })
+import { MaterialIcon } from "@/components/ui/MaterialIcon"
 
 export default function ChatPage() {
-  const newUI = isNewUIEnabled()
-  
-  // Always call hooks - they must be called unconditionally
   const lockedHeight = useLockedViewportHeight()
   const containerStyle = useMemo(() => {
     if (typeof lockedHeight !== "number") return undefined
     return { minHeight: lockedHeight, height: lockedHeight }
   }, [lockedHeight])
 
-  if (newUI) {
-    return <ChatPageNew />
-  }
-
-  // Original UI
+  const currentDate = new Date().toLocaleDateString('en-US', {
+    weekday: 'long',
+    month: 'long',
+    day: 'numeric'
+  })
 
   return (
     <Suspense fallback={null}>
       <div
-        className={`${inter.variable} font-sans min-h-screen h-screen relative overflow-hidden`}
+        className="font-sans min-h-screen h-screen relative overflow-hidden bg-[var(--hs-bg)] flex flex-col"
         style={containerStyle}
       >
-        <Link
-          href="/today"
-          className="absolute left-3 top-3 z-10 inline-flex items-center rounded-md border border-border/40 bg-card/20 px-3 py-1.5 text-xs text-foreground/80 backdrop-blur transition-opacity hover:opacity-90 focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-primary/60"
-        >
-          Home
-        </Link>
+        {/* Top App Bar - Headspace style */}
+        <header className="flex items-center px-4 py-3 justify-between sticky top-0 z-10 bg-[var(--hs-bg)]/95 backdrop-blur-md border-b border-[var(--hs-border-subtle)]">
+          <Link
+            href="/"
+            className="w-10 h-10 rounded-full flex items-center justify-center text-[var(--hs-text-secondary)] hover:bg-[var(--hs-surface)] transition-colors"
+            aria-label="Go back"
+          >
+            <MaterialIcon name="arrow_back" />
+          </Link>
+
+          <div className="flex-1 text-center">
+            <h2 className="text-base font-semibold text-[var(--hs-text-primary)]">
+              Journal
+            </h2>
+            <p className="text-xs text-[var(--hs-text-tertiary)]">
+              {currentDate}
+            </p>
+          </div>
+
+          <button
+            className="w-10 h-10 rounded-full flex items-center justify-center text-[var(--hs-text-secondary)] hover:bg-[var(--hs-surface)] transition-colors"
+            aria-label="More options"
+          >
+            <MaterialIcon name="more_vert" />
+          </button>
+        </header>
+
+        {/* Chat component - kept as-is for AI SDK compatibility */}
         <EtherealChat />
       </div>
     </Suspense>
