+++ components/home/CheckInSlots.tsx
@@ -4,16 +4,10 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import {
   type CheckInSlotState,
   useDailyCheckIns,
-  MORNING_START_HOUR,
-  EVENING_START_HOUR,
 } from '@/hooks/useDailyCheckIns'
-import { Button } from '@/components/ui/button'
 import { GuardedLink } from '@/components/common/GuardedLink'
-import { Skeleton } from '@/components/ui/skeleton'
-import { Badge } from '@/components/ui/badge'
-import { Alert, AlertDescription, AlertTitle } from '@/components/ui/alert'
-import { CHECK_IN_DRAFT_PREFIX } from '@/lib/check-ins/shared'
 import { cn } from '@/lib/utils'
+import { CHECK_IN_DRAFT_PREFIX } from '@/lib/check-ins/shared'
 
 interface DraftState {
   morning: boolean
@@ -25,7 +19,7 @@ interface CheckInSlotsProps {
 }
 
 export function CheckInSlots({ selectedDate = new Date() }: CheckInSlotsProps) {
-  const { isLoading, error, refetch, isViewingToday, targetDate, morning, evening, streak } =
+  const { isLoading, error, isViewingToday, targetDate, morning, evening } =
     useDailyCheckIns(selectedDate)
 
   const targetDateIso = useMemo(() => targetDate.toISOString().slice(0, 10), [targetDate])
@@ -65,38 +59,44 @@ export function CheckInSlots({ selectedDate = new Date() }: CheckInSlotsProps) {
   }, [updateDrafts])
 
   if (isLoading) {
-    return <CheckInSlotsSkeleton />
+    return (
+      <>
+        <div className="flex justify-between items-start gap-4 p-3 rounded-lg bg-gray-100/60 dark:bg-gray-700/40 animate-pulse">
+          <div className="flex flex-col gap-1 flex-1">
+            <div className="h-5 bg-gray-300 dark:bg-gray-600 rounded w-3/4" />
+            <div className="h-4 bg-gray-200 dark:bg-gray-700 rounded w-full" />
+          </div>
+          <div className="h-6 w-20 bg-gray-200 dark:bg-gray-700 rounded-full" />
+        </div>
+        <div className="flex justify-between items-start gap-4 p-3 rounded-lg text-gray-400 dark:text-gray-500 animate-pulse">
+          <div className="flex flex-col gap-1 flex-1">
+            <div className="h-5 bg-gray-300 dark:bg-gray-600 rounded w-3/4" />
+            <div className="h-4 bg-gray-200 dark:bg-gray-700 rounded w-full" />
+          </div>
+          <div className="h-6 w-20 bg-gray-200 dark:bg-gray-700 rounded-full" />
+        </div>
+      </>
+    )
   }
 
   if (error) {
     return (
-      <div className="col-span-2">
-        <Alert variant="destructive">
-          <AlertTitle>Unable to load check-ins</AlertTitle>
-          <AlertDescription>{error}</AlertDescription>
-        </Alert>
-        <Button variant="outline" className="mt-3" onClick={refetch}>
-          Try again
-        </Button>
+      <div className="col-span-2 text-red-500 text-center p-6 rounded-lg border border-red-500/20 bg-red-500/5">
+        <p>Unable to load check-ins: {error}</p>
       </div>
     )
   }
 
   return (
     <>
-      {isViewingToday && streak > 0 ? (
-        <div className="col-span-2">
-          <StreakBanner streak={streak} />
-        </div>
-      ) : null}
-      <SlotCard
+      <SlotCardNew
         variant="morning"
         state={morning}
         isViewingToday={isViewingToday}
         targetDate={targetDate}
         hasDraft={drafts.morning}
       />
-      <SlotCard
+      <SlotCardNew
         variant="evening"
         state={evening}
         isViewingToday={isViewingToday}
@@ -107,7 +107,7 @@ export function CheckInSlots({ selectedDate = new Date() }: CheckInSlotsProps) {
   )
 }
 
-interface SlotCardProps {
+interface SlotCardNewProps {
   variant: 'morning' | 'evening'
   state: CheckInSlotState
   isViewingToday: boolean
@@ -115,13 +115,11 @@ interface SlotCardProps {
   hasDraft: boolean
 }
 
-function SlotCard({ variant, state, isViewingToday, targetDate, hasDraft }: SlotCardProps) {
+function SlotCardNew({ variant, state, isViewingToday, targetDate, hasDraft }: SlotCardNewProps) {
   const href = variant === 'morning' ? '/check-in/morning' : '/check-in/evening'
   const title = variant === 'morning' ? 'Morning check-in' : 'Evening reflection'
 
   const statusLabel = getStatusLabel(state.status, hasDraft)
-  const statusVariant = getStatusBadgeVariant(state.status, hasDraft)
-
   const description = getStatusDescription({
     variant,
     state,
@@ -131,32 +129,64 @@ function SlotCard({ variant, state, isViewingToday, targetDate, hasDraft }: Slot
   })
 
   const showAction = state.status === 'available'
-  const actionLabel = hasDraft ? 'Resume' : 'Begin'
+  const isAvailable = state.status === 'available'
+  const isCompleted = state.status === 'completed'
+  const displayStatusLabel = isCompleted ? statusLabel : 'Not recorded'
+
+  const cardContent = (
+    <>
+      <div className="flex flex-col gap-1">
+        <p className={cn(
+          'text-lg font-bold',
+          isAvailable
+            ? 'text-gray-800 dark:text-gray-100'
+            : 'text-gray-400 dark:text-gray-500'
+        )}>
+          {title}
+        </p>
+        <p className={cn(
+          'text-sm',
+          isAvailable
+            ? 'text-gray-600 dark:text-gray-300'
+            : 'text-gray-400 dark:text-gray-500'
+        )}>
+          {description}
+        </p>
+      </div>
+      <div className={cn(
+        'px-3 py-1 text-xs font-medium rounded-full whitespace-nowrap',
+        'bg-gray-200 dark:bg-gray-600 text-gray-600 dark:text-gray-300'
+      )}>
+        {displayStatusLabel}
+      </div>
+    </>
+  )
+
+  if (showAction) {
+    return (
+      <GuardedLink href={href} className="block">
+        <div
+          data-slot={variant}
+          className={cn(
+            'flex justify-between items-start gap-4 p-3 rounded-lg cursor-pointer',
+            'bg-gray-100/60 dark:bg-gray-700/40 hover:bg-gray-200/60 dark:hover:bg-gray-700/60 transition-colors'
+          )}
+        >
+          {cardContent}
+        </div>
+      </GuardedLink>
+    )
+  }
 
   return (
     <div
       data-slot={variant}
       className={cn(
-        'col-span-2 md:col-span-1 flex h-full flex-col justify-between rounded-2xl border border-border bg-card/60 p-5 shadow-sm backdrop-blur transition hover:border-border/80',
+        'flex justify-between items-start gap-4 p-3 rounded-lg',
+        'text-gray-400 dark:text-gray-500'
       )}
     >
-      <div className="flex items-start justify-between gap-3">
-        <div>
-          <p className="text-xs font-medium uppercase tracking-wide text-muted-foreground">
-            {variant === 'morning' ? 'Morning' : 'Evening'}
-          </p>
-          <h2 className="mt-1 text-lg font-semibold">{title}</h2>
-        </div>
-        <Badge variant={statusVariant}>{statusLabel}</Badge>
-      </div>
-      <p className="mt-3 text-sm text-muted-foreground">{description}</p>
-      {showAction ? (
-        <GuardedLink href={href} className="mt-5 inline-flex w-full">
-          <Button className="w-full min-h-12 py-3" variant={hasDraft ? 'secondary' : 'default'}>
-            {actionLabel}
-          </Button>
-        </GuardedLink>
-      ) : null}
+      {cardContent}
     </div>
   )
 }
@@ -181,16 +211,6 @@ function getStatusLabel(status: CheckInSlotState['status'], hasDraft: boolean) {
   }
 }
 
-function getStatusBadgeVariant(
-  status: CheckInSlotState['status'],
-  hasDraft: boolean,
-): 'default' | 'secondary' | 'outline' {
-  if (status === 'completed') return 'default'
-  if (status === 'available' && hasDraft) return 'secondary'
-  if (status === 'available') return 'default'
-  return 'outline'
-}
-
 interface StatusDescriptionArgs {
   variant: 'morning' | 'evening'
   state: CheckInSlotState
@@ -200,7 +220,7 @@ interface StatusDescriptionArgs {
 }
 
 function getStatusDescription(args: StatusDescriptionArgs) {
-  const { variant, state, isViewingToday, targetDate, hasDraft } = args
+  const { variant, state, isViewingToday, targetDate } = args
 
   switch (state.status) {
     case 'completed':
@@ -208,16 +228,13 @@ function getStatusDescription(args: StatusDescriptionArgs) {
         ? 'Thanks for checking in today.'
         : 'This check-in was completed on this date.'
     case 'available':
-      if (hasDraft) {
-        return 'Pick up where you left off and finish this check-in when you are ready.'
-      }
       return variant === 'morning'
-        ? 'Ease into the day with a quick reflection.'
-        : 'Wind down the day with a gentle review.'
+        ? 'Start your day with a quick reflection.'
+        : 'Return this evening to reflect on your day.'
     case 'locked':
-      return `Evening reflections open later today at ${formatAvailableTime(state.availableAt, variant)}.`
+      return `Evening reflections open later today.`
     case 'upcoming':
-      return `Morning check-ins open at ${formatAvailableTime(state.availableAt, variant)}.`
+      return `Morning check-ins open soon.`
     case 'closed':
       return 'Come back tomorrow for a fresh start.'
     default:
@@ -226,72 +243,7 @@ function getStatusDescription(args: StatusDescriptionArgs) {
           ? 'Start your day with a quick reflection once the window opens.'
           : 'Return this evening to reflect on your day.'
       }
-      return `No check-in was recorded for ${formatDisplayDate(targetDate)}.`
+      return `No check-in was recorded for ${targetDate.toLocaleDateString('en-US', { month: 'short', day: 'numeric' })}.`
   }
 }
 
-function formatAvailableTime(time: string | null | undefined, variant: 'morning' | 'evening') {
-  if (time) {
-    const [hourString, minuteString] = time.split(':')
-    const hour = parseInt(hourString ?? '0', 10)
-    const minute = parseInt(minuteString ?? '0', 10)
-    return formatTimeLabel(hour, minute)
-  }
-  const fallbackHour = variant === 'morning' ? MORNING_START_HOUR : EVENING_START_HOUR
-  return formatTimeLabel(fallbackHour, 0)
-}
-
-function formatDisplayDate(date: Date) {
-  return date.toLocaleDateString('en-US', {
-    weekday: 'long',
-    month: 'short',
-    day: 'numeric',
-  })
-}
-
-function formatTimeLabel(hour24: number, minute: number) {
-  const hour = Number.isFinite(hour24) ? hour24 : 0
-  const minuteClamped = Number.isFinite(minute) ? Math.max(0, Math.min(59, minute)) : 0
-  const hour12 = (hour % 12 + 12) % 12 || 12
-  const suffix = hour >= 12 ? 'PM' : 'AM'
-  const minuteLabel = String(minuteClamped).padStart(2, '0')
-  return `${hour12}:${minuteLabel} ${suffix}`
-}
-
-function CheckInSlotsSkeleton() {
-  return (
-    <>
-      <div className="col-span-2 rounded-2xl border border-border/60 bg-card/40 p-5">
-        <Skeleton className="h-4 w-24" />
-        <Skeleton className="mt-2 h-6 w-1/2" />
-        <Skeleton className="mt-4 h-9 w-28" />
-      </div>
-      <div className="col-span-2 md:col-span-1 rounded-2xl border border-border/60 bg-card/40 p-5">
-        <Skeleton className="h-4 w-20" />
-        <Skeleton className="mt-2 h-5 w-3/4" />
-        <Skeleton className="mt-5 h-10 w-full" />
-      </div>
-      <div className="col-span-2 md:col-span-1 rounded-2xl border border-border/60 bg-card/40 p-5">
-        <Skeleton className="h-4 w-20" />
-        <Skeleton className="mt-2 h-5 w-3/4" />
-        <Skeleton className="mt-5 h-10 w-full" />
-      </div>
-    </>
-  )
-}
-
-function StreakBanner({ streak }: { streak: number }) {
-  return (
-    <div className="flex items-center justify-between rounded-2xl border border-border/60 bg-card/60 p-4 shadow-sm backdrop-blur">
-      <div>
-        <p className="text-xs font-medium uppercase tracking-wide text-muted-foreground">Momentum</p>
-        <p className="mt-1 text-sm text-muted-foreground">
-          You are on a {streak}-day check-in streak. Keep the rhythm going.
-        </p>
-      </div>
-      <div className="text-2xl" aria-hidden>
-        🔥
-      </div>
-    </div>
-  )
-}
